segregated IPC calls and App
